+++ selftest-nologger-tests/src/main/java/net/oneandone/httpselftest/example/selftest/SelftestConfiguration.java
@@ -1,0 +1,49 @@
+package net.oneandone.httpselftest.example.selftest;
+
+import java.util.function.Supplier;
+
+import javax.inject.Inject;
+
+import org.springframework.boot.actuate.endpoint.web.EndpointServlet;
+import org.springframework.boot.actuate.endpoint.web.annotation.ServletEndpoint;
+import org.springframework.boot.web.servlet.FilterRegistrationBean;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.core.env.Environment;
+
+import net.oneandone.httpselftest.servlet.SelftestMDCFilter;
+import net.oneandone.httpselftest.servlet.SelftestServlet;
+
+@Configuration
+public class SelftestConfiguration {
+
+    @Inject
+    Environment env;
+
+    @Bean
+    public SelftestEndpoint selftestEndpoint() {
+        return new SelftestEndpoint();
+    }
+
+    @Bean
+    public FilterRegistrationBean<SelftestMDCFilter> filterRegistration() {
+        FilterRegistrationBean<SelftestMDCFilter> reg = new FilterRegistrationBean<>();
+        reg.addUrlPatterns("/*");
+        reg.setFilter(new SelftestMDCFilter());
+        return reg;
+    }
+
+    @ServletEndpoint(id = "selftest")
+    public class SelftestEndpoint implements Supplier<EndpointServlet> {
+
+        @Override
+        public EndpointServlet get() {
+            return new EndpointServlet(NoLoggerSelftestServlet.class) //
+                    .withInitParameter(SelftestServlet.PROP_LOGGER, env.getProperty(SelftestServlet.PROP_LOGGER)) //
+                    .withInitParameter(SelftestServlet.PROP_OVERRIDE_PORT, env.getProperty(SelftestServlet.PROP_OVERRIDE_PORT)) //
+                    .withInitParameter(SelftestServlet.PROP_OVERRIDE_PATH, env.getProperty(SelftestServlet.PROP_OVERRIDE_PATH));
+        }
+
+    }
+
+}