+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 <!-- Categories: Added, Changed, Deprecated, Removed, Fixed, Security -->
 
 ## [Unreleased]
+### Changed
+- Make logback dependency optional. Logging framework is now configurable under 'selftest.logger', defaulting to logback.
 
 ## [0.1.0] - 2019-02-08
 ### Added